Rename package entity to domain
entity 패키지의 이름을 domain으로 변경했습니다.
api controller에서 결과값으로 반환하기 위한 Result 객체 등도 포함하기 위함입니다.
