+++ src/main/java/com/codessquad/qna/repository/AnswerRepository.java
@@ -1,6 +1,6 @@
 package com.codessquad.qna.repository;
 
-import com.codessquad.qna.entity.Answer;
+import com.codessquad.qna.domain.Answer;
 
 public interface AnswerRepository extends GenericRepository<Answer, Long> {
 }
