+++ src/main/java/com/codessquad/qna/service/UserService.java
@@ -1,6 +1,6 @@
 package com.codessquad.qna.service;
 
-import com.codessquad.qna.entity.User;
+import com.codessquad.qna.domain.User;
 import com.codessquad.qna.exception.AlreadyExistException;
 import com.codessquad.qna.exception.NotAuthorizedException;
 import com.codessquad.qna.exception.NotFoundException;
